refactor common file navigation code into singleton
