+++ src/gwt/src/org/rstudio/studio/client/RStudioGinModule.java
@@ -20,6 +20,7 @@ import com.google.gwt.user.client.ui.Widget;
 import com.google.inject.Singleton;
 import com.google.inject.name.Names;
 
+import org.rstudio.core.client.CodeNavigationListener;
 import org.rstudio.core.client.command.ApplicationCommandManager;
 import org.rstudio.core.client.command.EditorCommandManager;
 import org.rstudio.core.client.command.ShortcutViewer;
@@ -286,6 +287,7 @@ public class RStudioGinModule extends AbstractGinModule
       bind(ProjectTemplateRegistryProvider.class).in(Singleton.class);
       bind(PackageProvidedExtensions.class).asEagerSingleton();
       bind(JavaScriptEventHistory.class).asEagerSingleton();
+      bind(CodeNavigationListener.class).asEagerSingleton();
 
       bind(ApplicationView.class).to(ApplicationWindow.class)
             .in(Singleton.class) ;
